+++ DesktopManager/utils/audioUtils/IAudioPlayer.h
@@ -15,7 +15,7 @@ public:
     virtual void resume() = 0;
     virtual void pause() = 0;
     // 0-100
-    virtual void setVolume(uint32_t volume) = 0;
+    virtual void setVolume(uint8_t volume) = 0;
 };
 
 #endif // IAUDIOPLAY_H
